feat(common) 폴더 구조 변경. application.yaml 수정
