--- academic-service/src/main/java/com/green/academic/calendar/application/CalendarService.java
@@ -1,7 +1,0 @@
-package com.green.academic.calendar.application;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CalendarService {
-}
