+++ academic-service/src/main/java/com/green/academic/controller/AdminCalandarController.java
@@ -1,9 +1,9 @@
-package com.green.core.major.application;
+package com.green.academic.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/majors")
-public class MajorController {
+@RequestMapping("/api/admin/lectures")
+public class AdminCalandarController {
 }
