+++ academic-service/src/main/java/com/green/academic/service/CalandarService.java
@@ -1,0 +1,7 @@
+package com.green.academic.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CalandarService {
+}
