+++ core-service/src/main/java/com/green/core/controller/GradeController.java
@@ -1,9 +1,9 @@
-package com.green.core.lecture.application;
+package com.green.core.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/lectures")
-public class LectureController {
+@RequestMapping("/api/calendars")
+public class GradeController {
 }
