+++ core-service/src/main/java/com/green/core/entity/AcademicCalandarsCache.java
@@ -1,0 +1,4 @@
+package com.green.core.entity;
+
+public class AcademicCalandarsCache {
+}
