--- core-service/src/main/java/com/green/core/lecture/entity/Lectures.java
@@ -1,4 +1,0 @@
-package com.green.core.lecture.entity;
-
-public class Lectures {
-}
