+++ core-service/src/main/java/com/green/core/model/lecture/LectureCreateReq.java
@@ -1,0 +1,4 @@
+package com.green.core.model.lecture;
+
+public class LectureCreateReq {
+}
